backend/requirements.txt
# FastAPI and Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# AI and External APIs
groq==0.4.2
httpx==0.25.1
aiohttp==3.9.1

# Utilities
email-validator==2.1.0